Refactor `Action` into 'control/action'
