+++ src/control/action/action.go
@@ -1,0 +1,3 @@
+package action
+
+type Action = func()
