+++ src/control/editor/event_editor.go
@@ -1,4 +1,4 @@
-package control
+package editor
 
 import (
 	"strconv"
